Add HashTool
This is probably my last addition to the C++ version of this code,
at least if an Experimental ~~Rodeo~~ Rewrite doesn't end up panning
out.

This is a tool for generating the HashID values that the game uses.

Maybe later on I might try and see if I can write a multi-threaded
hash cracker, or better yet, a GPU powered one (no better way to
try and learn GPGPU then actually doing it!), but for now, this is
all you get.

Feel free to lavish in my argument parsing implementation. I just wanted
something simple and fairly easy to read, and this seems to work ok,
while still being platform-independent (no getopt, and I wouldn't need to
drag in additional code to reimplement getopt or such for Windows/Mac/...)